pyproject.toml
[tool.poetry]
name = "data-ingestion-trips-taken"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.3"
pandas = "^1.4.1"
SQLAlchemy = "^1.4.34"
redis = "^4.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
